common function

+++ src/test/java/com/github/nhathuyld/shop365/model/Product.java
@@ -3,12 +3,12 @@ package com.github.nhathuyld.shop365.model;
 public class Product {
 	public String name;
 	public String sku;
-	public int price;
-	public int pointshop;
-	public int shippingfee;
-	public int totalAmount;
-	public int pointEarn;
-	public int quantity;
+	public double price;
+	public double pointshop;
+	public double shippingfee;
+	public double totalAmount;
+	public double pointEarn;
+	public double quantity;
 	public String shippingName;
 
 }
